services/_template/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
pymysql==1.1.1
pika==1.3.2
requests==2.32.3
python-dotenv==1.0.1